backend/requirements.txt
Flask==2.3.3
   Flask-SQLAlchemy==3.0.5
   Flask-CORS==4.0.0
   PyJWT==2.8.0
   Flask-Bcrypt==1.0.1
   python-dotenv==1.0.0